Add a reporter based on Java's management libraries
This reports memory usage, thread counts and some primitive information
about GC. This allows us to expose a lot of the same information that
tools like VisualVM would be able to see.

Also a couple of minor improvements to error reporting, logging bad
requests and being more granular as to where we catch exceptions.